Sửa doanh thu

+++ src/main/java/service/impl/ThongKeServiceImpl.java
@@ -9,6 +9,7 @@ import custommodel.ThongKeSanPhamRespone;
 import java.awt.CardLayout;
 import java.awt.Color;
 import java.awt.Dimension;
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 import javax.swing.JPanel;
@@ -113,17 +114,17 @@ public class ThongKeServiceImpl implements ThongKeService {
         DefaultCategoryDataset dataset = new DefaultCategoryDataset();
         if (thang == 1 || thang == 3 || thang == 5 || thang == 7 || thang == 8 || thang == 10 || thang == 12) {
             for (int i = 1; i <= 31; i++) {
-                dataset.addValue(new Double(res.bieuDoDoanhThuMonth(i, thang, nam)), "Doanh số", i + "");
+                dataset.addValue(new BigDecimal(res.bieuDoDoanhThuMonth(i, thang, nam)), "Doanh số", i + "");
             }
         }
         if (thang == 2) {
             for (int i = 1; i <= 28; i++) {
-                dataset.addValue(new Double(res.bieuDoDoanhThuMonth(i, thang, nam)), "Doanh số", i + "");
+                dataset.addValue(new BigDecimal(res.bieuDoDoanhThuMonth(i, thang, nam)), "Doanh số", i + "");
             }
         }
         if (thang == 4 || thang == 6 || thang == 9 || thang == 11) {
             for (int i = 1; i <= 30; i++) {
-                dataset.addValue(new Double(res.bieuDoDoanhThuMonth(i, thang, nam)), "Doanh số", i + "");
+                dataset.addValue(new BigDecimal(res.bieuDoDoanhThuMonth(i, thang, nam)), "Doanh số", i + "");
             }
         }
         JFreeChart barChart = ChartFactory.createBarChart(
@@ -146,7 +147,7 @@ public class ThongKeServiceImpl implements ThongKeService {
     public void bieuDoDoanhThuYear(int thang, int nam, JPanel jpnItem) {
         DefaultCategoryDataset dataset = new DefaultCategoryDataset();
         for (int i = 1; i <= 12; i++) {
-            dataset.addValue(new Double(res.bieuDoDoanhThuYear(i, nam)), "Doanh số", i + "");
+            dataset.addValue(new BigDecimal(res.bieuDoDoanhThuYear(i, nam)), "Doanh số", i + "");
         }
         JFreeChart barChart = ChartFactory.createBarChart(
                 "Biểu đồ thống kê doanh thu năm " + nam + "".toUpperCase(),
